backend done, is admin check is working and additional adapter/routes…

+++ routes/sports.js
@@ -29,11 +29,18 @@ sportsRouter.get('/:sportId',async(req,res,next)=>{
 sportsRouter.post('/',authRequired,async(req,res,next)=>{
     try {
         const {name,description} = req.body;
-        const sport = await createSport({name,description});
-        //need to check if they are an admin
-        res.send({
-            sport
-        });
+        if (req.user.is_admin){
+            const sport = await createSport({name,description});
+            res.send({
+                sport
+            });
+        }
+        else{
+            next({
+                name: 'UserNotAdminError',
+                message: 'The user is not an admin and cannot perform this task'
+            })
+        }
     } catch (error) {
         next(error);
     }
@@ -43,26 +50,39 @@ sportsRouter.patch('/:sportId',authRequired,async(req,res,next)=>{
     try {
         const {sportId} = req.params;
         const {name,description} = req.body;
-        const sport = await getSportById(sportId);
-        // need to check if they are an admin
-        const updatedSport = await updateSport(sportId,name,description);
-        res.send({
-            updatedSport
-        })
+        if (req.user.is_admin){
+            const updatedSport = await updateSport(sportId,name,description);
+            res.send({
+                updatedSport
+            })
+        }
+        else{
+            next({
+                name: 'UserNotAdminError',
+                message: 'The user is not an admin and cannot perform this task'
+            })
+        }
     } catch (error) {
         next(error);
     }
 })
 
 sportsRouter.delete('/:sportId',authRequired,async(req,res,next)=>{
+    console.log(req.user)
     try {
         const {sportId} = req.params;
-        const sport = await getSportById(sportId);
-        // need to check if they are an admin
-        const deletedSport = await destroySport(sportId);
-        res.send({
-            deletedSport
-        })
+        if (req.user.is_admin){
+            const deletedSport = await destroySport(sportId);
+            res.send({
+                deletedSport
+            })
+        }
+        else{
+            next({
+                name: 'UserNotAdminError',
+                message: 'The user is not an admin and cannot perform this task'
+            })
+        }
     } catch (error) {
         next(error);
     }
